requirements.txt
camelot-py[base]
PyPDF2<3.0
opencv-python-headless
deta
python-dotenv
requests
prophet
dropbox
pymsteams
matplotlib
tqdm
XlsxWriter
neuralprophet